test(utils): isolate utils tests
