+++ src/utils/generateUUID.spec.ts
@@ -1,0 +1,11 @@
+import { generateUUID } from './generateUUID';
+
+describe('Generate UUID', () => {
+	it('should generate a UUID', async () => {
+		const uuid = generateUUID();
+
+		expect(uuid).toBeDefined();
+		expect(typeof uuid).toBe('string');
+		expect(uuid.length).toBe(36);
+	});
+});
